+++ .gitignore
@@ -2,3 +2,4 @@ bot_token.txt
 path/*
 db-info/config.yml
 db-info/db-wallet
+commands/__pycache__/*
