+++ commands/__init__.py
@@ -1,0 +1,17 @@
+from .character import create_character, delete_character, delete_all_characters, edit_character, allow_character, view_character
+from .db import init, load_characters_from_message, export_characters_manual, create_oracle_connection
+from .messaging import speak_as_character
+
+__all__ = [
+    'create_character',
+    'delete_character',
+    'delete_all_characters',
+    'edit_character',
+    'allow_character',
+    'view_character',
+    'init',
+    'load_characters_from_message',
+    'export_characters_manual',
+    'speak_as_character',
+    'create_oracle_connection'
+]
